Replaced broken link, not archived due robos.txt
The webpage http://web.archive.org/web/https://landfill.bugzilla.org/bugzilla-tip/page.cgi?id=bug-writing.html,
cannot arquive the link due its robots configuration, so as the page is off-line, the link is completely broken.
On this was added two new links, hopefully to be good enough replacers for the original one.

The first link is equivalent to the original broken. Actually, inside it, there is a link to the original one, which seens offline,
until their sever update some time in the future.